feat: load full 12KB of the kernel from disk when bootloading
Our kernel has now exceeded 3 disk sectors,
which means that we need to load more of the kernel into memory from disk in the bootloader. We were only loading three sectors to get us started. Luckily we already have the code to read from an arbitrary number of sectors in `os_read_sectors`. We can just borrow that code.